Associate content with kinds
A kind is more specific content (either an episode or chapter). The main purpose of this is to support list entries for titles with a certain amount of content which has "essence". For example, Kino's Journey (2003) has 13 episodes which span ~24 minutes each (differing usually in seconds). Most lists allow users to say how much content they've watched, but it's usually just an amount (i.e. a number), which doesn't allow fine-tuning (e.g. only watching a subset of the content in a non-sequential order). The "essence" here is the prime detail of the content (for episodes, how long in seconds / duration; for chapters, how many pages), which would allow further fine-tuning with progress (e.g. 68% done with chapter X).