Use `.sheet(item:onDismiss:content:)` to present title cover sheet
The primary purpose of this is to push creating child view contexts from inside the view (which is a bad idea, since it entails side effects) to outside it in the button action. Since sheet supports an onDismiss argument, coordinating the change has been moved to it as well, as opposed to relying on `.onDisappear(action:)`.

I was initially surprised that NSManagedContext supported conforming to Identifiable without extra work.