Add back create entry button
I'm currently thinking about how to organize entries since they often don't have much identifying information. Specifically, whether or not users should be bound to creating entries for a title or directly on content instead. The former has the advantage of having a clear timeline of all the content that can be entries, but the latter has the advantage of potentially being groupable with anything that can have an entries (e.g. across titles). I think the latter would make more sense, but I currently don't know how to design the interface. The next commit will likely group content on titles for an idea.