Create schema for list plans
A list plan is a set of content. It'll be used as a replacement for the common "Create List Entry" interface standard on many list platforms where you can only record the details of a watch/read (usually only one) linearly, for that title. A list plan would make it simple for a user to specify which contents they plan on consuming while allowing them to go cross-title (e.g. binge watching a series).